hello-world.S: Fix missing stack size on setup
Commit e92abac removed an `li` statement that would have set up our
stack to be of size 0x2000 at the end of the payload, in combination
with the next statement.

Without explicitly loading immedate 0x2000, we don't actually know how
large the stack size would be. This depends on the previous value of the
register and can very well be 0, thus overwriting the executable (or at
least its BSS and DATA sections) on push operations.