Adapt Sv32 paging terminology from specification
According to the RISC-V Privileged Architecture specification, Sv32
paging is implemented as a radix tree, where the uppermost node is
called the `root page table`. Due to its variable height, it seems like
there is no explicit terminology for leaf page tables, however, page
table entries (PTEs) that do not point to another page table level but
to a physical page number instead are reffered to as leaf PTEs.