kmap_kernel_upper_half: Fix invalid stack page mapping
initial_stack_start returns the initial state of the sp. However,
because RISC-V psABI requires a decreasing full stack, sp initially does
not point to the last address in the last stack page but rather to first
address of the first page that can be allocated using kpalloc(). Thus,
the stack has been falsely mapped to the first allocated page which
happens to be the mid page table for 0x80000000 for 2MiB.
Effectively, after moving the stack pointer upwards to the upper half,
the stack created new invalid page table entries that were problematic.