perform_initial_ctxt_switch: Jump to upper half mirror of switch_to_user_pt
switch_to_user_pt expects to be in the upper half of the virtual address
space. Thus, we need to calculate the offset of the function relative to
the trap handler page and jump to the upper mirror of the trap handler
page plus the offset.
If we are entering from a trap, everything is already fine as the trap
handler has been set to the upper half beforehand.