Kernel kernel_environ_init: Perform initial state assertions
Use assertions to assume a well-defined initial state. As of right now,
the alignments of the trap handlers as well as the kernel page table are
checked. Unfortunately, it is not possible to check these alignments at
compile-time. It may be possible to check alignment on linker-time, but
assertions in the linker script feel like they are out of place.