MMU: Add kfree_page_table
Due to its radix tree nature, to destruct the page table and reclaim all
mapped pages and radix tree nodes, we need to iterate over all PTEs and
nodes.
While it would have been possible to implement this in a cleaner
recursive fashion, we have decided against this because the loops are
strictly bounded (a recursive function with a bound function argument
would work, too). The recursive function would have to check whether a
valid PTE is a node (recursive call and remove node) or a leaf (just
remove).