Move kmalloc to library OS glue source
kmalloc is not a system call but rather brk and malloc combined. Also,
due to kmalloc being used only in the library OS variant (the kernel
variant actually implemented brk in trap.c), it is best to move it from
the general system function set to the library OS glue.