mmu: Add kstrlcpy_from_vspace for string copies from user space
Unfortunately, it is rather unsafe to directly use a userspace pointer
that has been converted to kernel space addresses. We blindly trust a
user buffer and feed it to kopen. This is especially problematic if the
buffer crosses page boundaries as this does not necessarily mean that
both mapped pages are adjacent, so a simple pointer increment of
translated addresses is not possible.
kstrlcpy_from_vspace acts similar to down_load_string in selfie: It
copies the data word-wise. In each loop iteration, we translate the
address anew and by this way ensure that we are on the actually
corresponding kernel space address even if mapped pages are not
adjacent.
Of course it could be more efficient to calculate page boundaries and do
the copy page-wise, but this approach will suffice for now.