kstrlcpy_from_vspace: Return 0 if buffer vaddr is invalid or not mapped
In the case of a userspace pointer that is not a valid Sv39 vaddr or is
not mapped to userspace (if the buffer hosts valid contents, it must be
initialized and thus be mapped), return 0 as we were not able to
complete the copying process completely. A buffer that has been
partially copied is of no use, after all.

Note, however, that we do not null the destination buffer.