Grader: Set number of assignment arguments to '?'
argparse uses quantifiers +, * and ? like regex. In order to replicate
the previous behavior of the argument parser, the star quantifier has
been used and the amount of parameters has been limited by
parse_assignment. However, actually the argument is either not present
or specified once, which matches up with the question mark quantifier.

Thus, parse_assignment does only require a single string instead of a
list.