grader: Fix negative test cases of assembler-parser
The function lib.checks.check_execution determines how it should check
selfie for expected behavior using the `success_criteria` parameter. The
function does check for the parameter's type and if a boolean is
provided, the grader just checks whether execution shall
succeed (exit code == 0) or not (exit code != 0).
The actual value of success_criteria is not relevant in this case,
though.

As we expect the negative tests `invalid-argument.s`,
`missing-instruction.s` and `missing-literal.s` to fail, we need to
explicitly set `should_succeed` to false, as `success_criteria` has
default value `True` (of type `bool`)