Machine: Clean opensbi build directory after selfie-opensbi.elf build
OpenSBI's makefile target `all` declares prerequisites on the
fw_payload.bin file of the specified platform, which itself depends on
the fw_payload.elf file that we move out after the build has finished.
As the fw_payload.bin file still exists, make does not look further and
does not rebuild fw_payload.elf.

Instead, remove the whole firmware/ subdirectory of the platform's build
output directory. This is preferrable to a `make clean` because this
would also clean shared build artifacts.

The relevant lines of this behavior are (in OpenSBI v0.7):
- opensbi/firmware/objects.mk:30
- opensbi/Makefile:134
- opensbi/Makefile:300
- opensbi/Makefile:304