machine: Isolate OpenSBI binary build based on build profile
When enabling concurrent jobs in Make, the build of both build profiles
`library` and `kernel` on the same board may clash when they are
executed at the same time because the build artifacts were placed at the
same location.

This commit isolates the build of each build profile into a separate
OpenSBI build directory to prevent this. There is no need to isolate
based on the board as OpenSBI's build system already does this for us.