Grader: Don't re-assemble binary on Mipster in 'self-assembler' (fixes #229)
Currently, the `self-assembler` check causes a timeout for some
implementations. This is due to the assemble and disassemble step taking
rather long.

This commit instead emits two binaries. The first is generated by the
host's selfie by compiling, disassembling and re-assembling `selfie.c`.
The resulting binary is then run by Mipster to compile the source again
and emit the binary. If the assemble-step succeeded, the re-assembled
binary should emit itself and thus both binaries should be equivalent.