Grader: Fix assignment 'self-assembler' passing without impl
@fischer-martin informed me that the change made by #229 introduced a
bug.
The assignment will falsely succeed when support for the `-a` flag
has been implemented, but no changes to global `binary` have
been made. This is due to the previously compiled and disassembled
binary still being in memory.

This PR splits the first execution check into a preparation step that
compiles and disassembles `selfie.c` and another check that does the
assemble step, binary output, mipster compilation and binary output.