Grader: Don't fail assignment if it contains mandatory checks only
The grader does naturally not include mandatory checks into the
calculation of a student's grade. If there were no positive checks, the
grader will explicitly set the amount of `passed` tests to 0%.
Furthermore, it will add string `you have not passed at least one
positive test` to the reasons grade 5 was given.
Unfortunately, if an assignment contains mandatory tests only (important
for all-or-nothing assignments), the grader mistakenly sets 0% passed
rate. This is due to mandatory tests not being added to the check
counters. As there are no non-mandatory tests, there cannot be any
passed non-mandatory tests.
This commit adds an check whether `number_of_tests` is zero (either
mandatory tests only or no tests at all). In this case, the rate of
passed tests is explicitly set to 100%.