Grader: Allow SIGINT to be handled when waiting for a subprocess (#144)
When forking a child process, both the parent and the children reside in
the same process group. Signals like SIGINT that directed to a process
group will be distributed to all processes in the group. However, the
grader did not receive SIGINT when issuing [Ctrl]+[C] in a hosting
terminal. This is due to to the grader using `system` to start the child
processes.

As per the man-pages of `system` for
[FreeBSD](https://www.unix.com/man-page/freebsd/3/system/),
[macOS](https://developer.apple.com/library/archive/documentation/System/Conceptual/ManPages_iPhoneOS/man3/system.3.html)
and [Linux](https://linux.die.net/man/3/system), signal `SIGINT` will be
ignored while the application waits for the child process to finish.
Unfortunately, the return value of Python's `os.system` is
platform-dependent.

This commit instead uses `subprocess.run` which under the hood calls an
`execve`-like function. SIGINT will not be masked by this and will be
received by both the child `git` processes and the parent grader
process.