Correct cache hit counter in do_sd()
Since do_sd() is used to emulate a SD instruction, it should produce only one memory access and thus at most one
cash hit. load_virtual_memory() could result in a hit, or a miss, but store_virtual_memory() would definitely result
in a cache hit. This increases the number of potential cache hits to two, which is not the intended behavior.