Grader: Do not pass `thread` checks shared-(data|heap) if unimplemented
The checks `shared-heap` and `shared-data` erroneously passed when the
pthread syscalls were added, but not completely implemented. If the
return value of `pthread_create` happens to be 0, the main thread enters
the branch meant for the created thread, and does perform the operations
that pass the test.

This PR additionally uses the exit code to determine the check's exit
code. If the assignment has not been implemented, the test will fail due
to the missing thread's exit code.