machine: Generalize va_printf for multiple put operations
In order to prepare for sprintf, generalize the functionality of
va_printf in such a way that the caller has to specify a function that
handles putting string buffers (e.g. console_puts for printf or a simple
strncpy for sprintf)

While this implementation does not allow for specific optimizations, it
prevents code duplications.