Machine: Generalize RISC-U binary rule creation
As the general pattern to compile a RISC-U binary for packaging is the
same, use a Makefile macro to generate explicit build rules.
Unfortunately, it is not possible to create a general pattern rule with
an explicitly given source path (in the likes of
`%.m: %.c: ...` plus `$(BUILD_DIR)/bin.m: $(SRC_DIR)/bin.c`).

The generated rules contain a order-only prerequisite on the selfie
binary. An explicit rule to build selfie was necessary, which merely
delegates build to the root-level Makefile. If Selfie will ever have
dependencies on other source files, these dependencies would need to be
reflected or the recipe would need to be set as `.PHONY`