Fix merging of shared symbolic memory in monster.
This fixes the update process of shared portions of the symbolic memory
when contexts are merged. It is only valid to un-share the topmost
fragment of a shared memory when a context is merged with the context
that it was originally cloned from. This is mostly ensured by the order
in which merging happens, but no longer holds true in the presence of
multiple return statements within a function body.

The newly added example demonstrates this issue and causes the bounded
loop to erroneously become unbounded due to the bogus un-sharing.

Also note that un-sharing of shared portions can be thought of as an
optimization. It won't affect correctness if we miss some un-sharing
opportunities. Overzealous un-sharing however will cause issues as the
example demonstrates.