selfie: Add write_synced
Currently, selfie directly uses the host's `write` function when implementing `SYS_WRITE` for higher bootlevels. This however causes issues with glibc buffering when these instances attempt to write to FD 1, causing output to appear out of order to the grader. As such, a new function `write_synced` is added which properly handles the case of writing to console for the highest bootlevel, by using `printf` (which is properly synced) to do so.

This fixes the issue addressed in both #337 and #338.